+++ src/features/XIT/CONTS/PartnerLink.vue
@@ -15,5 +15,20 @@ defineProps<{ contract: PrunApi.Contract }>();
   <PrunLink v-else-if="contract.partner.name" :command="`CO ${contract.partner.code}`">
     {{ contract.partner.name }}
   </PrunLink>
-  <PrunLink v-else :command="`CO ${contract.partner.code}`" />
+  <PrunLink v-else-if="contract.partner.code" :command="`CO ${contract.partner.code}`" />
+  <div
+    v-else-if="contract.partner.currency"
+    data-tooltip="Refined PrUn is unable to fetch government information. Check the contract info for more details."
+    data-tooltip-position="down"
+    :class="$style.overrideTooltipStyle">
+    Planetary Government
+  </div>
+  <div v-else>Unknown</div>
 </template>
+
+<style module>
+.overrideTooltipStyle {
+  display: block;
+  padding: 0;
+}
+</style>